Update README and communication pool service
